Add use of target in GetBlocks message
This lets the ChainSynchronizer know when to stop asking
for new blocks, as we dont want to keep asking while we
are currently syncing. A zero hash can still be used
to simply ask for everything up to the tip of the chain,
as it will only cause an error which is logged, and wont
break anything.