Rename Extractor to Cache
The reason for this is that the Stakers struct
can already extract committees by itself. The Extractor is merely just
a cache, saving the client the hassle of having to re-extract
the committee for each event.
Additionally, the Stakers struct has been removed from the Cache.
It is now passed into the required methods, to avoid needing to
declare methods to adjust state on the interface types.