Unify Stakers and Provisioners
This is done because the Chain will need to use
Provisioners for certificate verification, so it makes no sense to
distinguish between the struct that consensus components get vs the one in the chain.
Additionally, Provisioners already exposes a TotalWeight method. The Stakers only purpose
was to have the total weight cached.