Refactor `idXor` and `collapseDistance`
- `Peer` distance is expressed now as `uint16` to hold the
max range value `128` unsoported on the `byte` type.
- Instead of using math logarithms, we work with bitwise ops.

- `collapseDistance` has been ranamed as `classifyDistance` and
optimized since now only counts the ammount of `1` of a byte that
comes from an XOR op.
- `idXor` just performs the XOR between the two `Peer-ID` arrays
and computes the number of different bits on the result, using this
counter as the result for the distance classification value.
NOTE that this value has the range (0 - 128] to match the `L-buckets`.