Add destPort to `Packet` headers in the protocol
Since the node that receives a ping, only knows the port
where the packet was sent, it does not know on which port
the sender is listening for the UDP response. And since it
is not possible to writte to a port that you are listening,
we need to send the listeningPort in the headers.

This also allows to update it.