Implement PeerSort struct
PeerSort struct aims to be a helper data
structure that allows the implementation
to order the PeerList of a Node in terms of
XOR-distance in respect to a certain Peer.

The struct has the same fields as the `Peer`
struct plus a `xorMyPeer` field which represents
the XOR-distance of this specific `Peer` in respect
to another.