Implement methods for FIND_NODES usage.
- Implemented sort.Interface for `ByXORDist`
that allows to sort slices of Peers by it's
XOR-distance.
- getPeerSortDist allows to get a list of
`PeerSort` items ready to be sorted by it's
XOR-distance in respect to some `Peer`.
- `getXClosestPeersTo` allows to get the selected
number of peers that is closer to the specified Peer.

Also implemented `FIND_NODES` message sender
function. Which gets the `alpha` closest nodes
to the Peer that runs the node and sends them
a `FIND_NODES` packet.