Refactor packetProcessing methodology.
Instead of spawning gorutines for each new
connection, we just add the received packet
to the circularqueue and the consumer will
take care of them.

The `Put` method sends a signal to the consumer
to awake and start consumming and processing
packets from the queue so we are not constantly
waiting.