implement GetCandidate, and send a round update with a round thats two above the block height, and remove the Seed and Hash fields
with the implementation of the correct certificate generation procedure, an accepted block height is always two behind the current round. furthermore, this accepted block seed and hash are not relevant to the consensus, so we strip them from the round update. the consensus coordinator will instead take care of getting them from an intermediate block