Implement basic skeleton for WaitGroups
Since `Sleep` was not an option as discussed in
 #141, the waitGroup has been implemented in order to
make the program itself trigger the protocol methods
once a packet type in concrete is received.

Also, added state-vars in order to force the
`packetProcessor` gorutine to just activate the
waitGroup if a bootstrapping or network discovery
process is in progress.

Now we need to add timeouts in order to prevent the
app to get stucked at any protocol method execution.