Refactor GetCandidate procedure
This commit introduces the candidate.Requestor,
which can be contacted over the RPCBus in order to
get a Candidate message from the network. This removes
the GetCandidate procedure from the chain entirely.
This commit also fixes an oversight in the first step reduction,
which would allow it to return on the first empty message
that it sees.