Create consensus integration testbed
The testbed aims to test the stability of the consensus
and chain working together, without using the network at all.
This is achieved by creating lots of stripped-down nodes
which all share the same event bus, and by routing gossip
messages directly back into the event bus.

This testbed is currently quite simple, but should prove
extensible in the future for various testing purposes.

Fixes #765