Change republishing to be in line with protocol
Since messages are only supposed to be
republished when the node is on the same state as the message,
we are now gossiping them directly out of the consensus components
after they are sufficiently verified. This also, currently, makes
the Republisher obsolete, but the code will be kept in case of
utility in future scenarios.

Fixes #760, fixes #758