Start consensus only after all nodes have been initialized
This fixes a bug, where some nodes would not
receive all messages properly, and cause TestConsensus to fail
over a panic when attempting to retrieve a candidate block
from the network. The panic occurs because there are no
CandidateBrokers initialized in this test, and thus the request
message falls on deaf ears. The upward propagation of the error
ultimately leads to a panic.

Fixes #866