Buffer sends on CatchBlockChan
This prevents blocks from being dropped. In order to
avoid halting on errors that get stuck in the channel,
errors are now to include blocks as well, so that the
recipient can determine whether or not it is ok to
halt on that particular error, based on block height.

Fixes #876