Add parsing transactions
This is achieved by replicating the structures in `phoenix-core`,
ensuring they are deserialized in *exactly* the same way as they are
serialized. This also enables us to write code to hash transactions on
this side of the code, allowing us to pass all information the explorer
needs to function properly.