Remove CommitteeSize function
CommitteeSize returned the minimum between the number of provisioners at a specific round
and MaxCommitteeSize (1 or 64, depending on the step). This was logically wrong, since
the 'size' of a committee does not represent, nor depend on the number of provisioners.

With this change, committees in reduction and agreement phases will always have size 64 'ConsensusCommitteeSize',
regardless of the number of available provisioners.